repositioning of switch layout button
